Se agregan los modulos que necesitaremos para hacer el modulo de cita, este es solo un avance

+++ managementApp/src/main/java/com/managementApp/repository/DireccionRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface DireccionRepository extends JpaRepository<Direccion, Integer> {
+public interface DireccionRepository extends JpaRepository<Direccion, Long> {
     Optional<Direccion> findByCalleAndColoniaAndNumeroExteriorAndNumeroInteriorAndCodigoPostal(
             String calle,
             String colonia,
