+++ managementApp/src/main/java/com/managementApp/repository/PersonaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PersonaRepository extends JpaRepository<Persona, Integer> {
+public interface PersonaRepository extends JpaRepository<Persona, Long> {
     Optional<Persona> findByNombreAndPrimerApeAndSegundoApe(
             String nombre,
             String primerApe,
